refactor api call to a service class
